feat: add principal video in list
